pyproject.toml
[tool.poetry]
name = "drosophila-brain"
version = "0.1.0"
description = ""
authors = ["lenafiedor <lenafiedor@o2.pl>"]
readme = "README.md"
packages = [{include = "drosophila_brain"}]

[tool.poetry.dependencies]
python = "^3.10"
neuprint-python = "^0.4.26"
selenium = "^4.21.0"
bokeh = "^3.4.1"
asciitree = "^0.3.3"
attrs = "^23.2.0"
certifi = "^2024.6.2"
cffi = "^1.16.0"
charset-normalizer = "^3.3.2"
colorama = "^0.4.6"
contourpy = "^1.2.1"
exceptiongroup = "^1.2.1"
h11 = "^0.14.0"
idna = "^3.7"
jinja2 = "^3.1.4"
markupsafe = "^2.1.5"
networkx = "^3.3"
numpy = "^2.0.0"
outcome = "^1.3.0.post0"
packaging = "^24.1"
pandas = "^2.2.2"
pillow = "^10.3.0"
pycparser = "^2.22"
pysocks = "^1.7.1"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
pyyaml = "^6.0.1"
requests = "^2.32.3"
scipy = "^1.13.1"
six = "^1.16.0"
sniffio = "^1.3.1"
sortedcontainers = "^2.4.0"
tornado = "^6.4.1"
tqdm = "^4.66.4"
trio-websocket = "^0.11.1"
trio = "^0.25.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.1"
ujson = "^5.10.0"
urllib3 = "^2.2.1"
wsproto = "^1.2.0"
xyzservices = "^2024.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
